clippy: module inception: move actor::actor into actor
